增加数据的权限控制

+++ src/main/java/com/ngtesting/platform/service/impl/CaseAttachmentServiceImpl.java
@@ -21,21 +21,30 @@ public class CaseAttachmentServiceImpl extends BaseServiceImpl implements CaseAt
     CaseDao caseDao;
 
     @Override
-    public void uploadAttachmentPers(Integer caseId, String name, String path, TstUser user) {
+    public Boolean save(Integer caseId, String name, String path, TstUser user) {
+        TstCase testCase = caseDao.get(caseId, user.getDefaultPrjId());
+        if (testCase == null) {
+            return false;
+        }
+
         TstCaseAttachment attach = new TstCaseAttachment(name, path, caseId, user.getId());
         caseAttachmentDao.save(attach);
-
-        TstCase testCase = caseDao.get(caseId, null);
         caseHistoryService.saveHistory(user, Constant.CaseAct.upload_attachment, testCase, name);
+        return true;
     }
 
     @Override
-    public void removeAttachmentPers(Integer id, TstUser user) {
-        caseAttachmentDao.delete(id);
-
+    public Boolean delete(Integer id, TstUser user) {
         TstCaseAttachment attach = caseAttachmentDao.get(id);
-        TstCase testCase = caseDao.get(attach.getCaseId(), null);
+        TstCase testCase = caseDao.get(attach.getCaseId(), user.getDefaultPrjId());
+        if (testCase == null) {
+            return false;
+        }
+
+        caseAttachmentDao.delete(id);
         caseHistoryService.saveHistory(user, Constant.CaseAct.delete_attachment, testCase, attach.getName());
+
+        return true;
     }
 
 }
